chore: renaming and sorting some functions
